% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{matriz}
\alias{matriz}
\title{Build the restriction matrix}
\usage{
matriz(n)
}
\arguments{
\item{n}{An integer.}
}
\value{
A (2n+1 X n^2 + n) matrix.
}
\description{
Build the restriction matrix
}
\examples{
matriz(3)
matriz(2)
}

